// src/interfaces/Config.ts
export interface ConfigConstant {
  JWT_PRIVATE_KEY: string;
  DB: string;
  DEV_INSTANCEURL: string;
  TEST_INSTANCEURL: string;
  PROD_INSTANCEURL: string;
  GTMVERSION: string;
  CURRENT_INSTANCE: string;
  IS_CLIENT: string;
  SCHEMA_URI: string;
  CLIENT_TOKEN: string;
  SUBSCIPTION: string[];
}